fix compiler warning in dnsreplay.cc

+++ pdns/dnsreplay.cc
@@ -310,7 +310,7 @@ uint64_t countLessThan(unsigned int msec)
 void emitFlightTimes()
 {
   uint64_t totals = countLessThan(flightTimes.size());
-  unsigned int limits[]={1, 2, 3, 4, 5, 10, 20, 30, 40, 50, 100, 200, 500, 1000, flightTimes.size()};
+  unsigned int limits[]={1, 2, 3, 4, 5, 10, 20, 30, 40, 50, 100, 200, 500, 1000, (unsigned int) flightTimes.size()};
   uint64_t sofar=0;
   cout.setf(std::ios::fixed);
   cout.precision(2);
